requirements.txt
mysql-connector-python
python-dotenv
Flask